Revert "Customer address"
This reverts commit 85e36c2d9905712b5d01356d95333836c67f4473.
