+++ CarEshipApp/app/src/main/java/com/example/careshipapp/CustomerAddress.java
@@ -1,0 +1,35 @@
+package com.example.careshipapp;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+
+public class CustomerAddress extends AppCompatActivity {
+
+    EditText postAddress, zipCode;
+    Button createAccountButton;
+    DBHelperClass database;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.customer_address);
+
+        database = new DBHelperClass(this);
+
+        createAccountButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+            }
+        });
+    }
+        }
+
+
