+++ CarEshipApp/app/src/main/java/com/example/careshipapp/MainActivityCreateAccount.java
@@ -4,7 +4,6 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Patterns;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
@@ -13,7 +12,7 @@ import com.google.android.material.button.MaterialButton;
 
 public class MainActivityCreateAccount extends AppCompatActivity {
 
-    EditText username, createPassword, passwordReentry, postAddress, zipCode;
+    EditText username, createPassword, passwordReentry;
     MaterialButton createAccountButton;
     DBHelperClass database;
 
@@ -25,8 +24,6 @@ public class MainActivityCreateAccount extends AppCompatActivity {
         username = (EditText) findViewById(R.id.Username);
         createPassword = (EditText) findViewById(R.id.CreatePassword);
         passwordReentry = (EditText) findViewById(R.id.passwordReentry);
-        postAddress = (EditText) findViewById(R.id.postAddress);
-        zipCode = (EditText) findViewById(R.id.zipCode);
         createAccountButton = (MaterialButton) findViewById(R.id.CreateAccountButton);
         database = new DBHelperClass(this);
 
@@ -36,39 +33,28 @@ public class MainActivityCreateAccount extends AppCompatActivity {
                 String email = username.getText().toString();
                 String passwrd = createPassword.getText().toString();
                 String repeatPassword = passwordReentry.getText().toString();
-                String address = postAddress.getText().toString();
-                String code = zipCode.getText().toString();
 
-                if(email.isEmpty() || passwrd.isEmpty() || repeatPassword.isEmpty() || code.isEmpty() || address.isEmpty()){
+                if(email.isEmpty() || passwrd.isEmpty() || repeatPassword.isEmpty()){
                     Toast.makeText(MainActivityCreateAccount.this, "Please enter non-empty values.", Toast.LENGTH_SHORT).show();
-                }
-                else if(!Patterns.EMAIL_ADDRESS.matcher(email).matches()){
-                    Toast.makeText(MainActivityCreateAccount.this,"please enter a valid email.",Toast.LENGTH_SHORT).show();
-                }
-                else if (zipCode.length() != 5)
-                {
-                    zipCode.setError("Please enter your 5 digit post code");
-                    zipCode.requestFocus();
                 }
                 else if(passwrd.equals(repeatPassword)) {
                     Boolean userCheck = database.userExistsCheck(email, passwrd);
                     Boolean usernameCheck = database.usernameExistsCheck(email);
 
                     if(userCheck == false && usernameCheck == false){
-                        database.insertData(email, passwrd,address, code);
-                        }
+                        database.insertData(email, passwrd);
                         Toast.makeText(MainActivityCreateAccount.this, "Account created successfully.", Toast.LENGTH_SHORT).show();
                         Intent intent = new Intent(MainActivityCreateAccount.this, MainActivityLoginPage.class);
                         startActivity(intent);
-                    }else {
-                    Toast.makeText(MainActivityCreateAccount.this, "Username or/and password are taken, please try again.", Toast.LENGTH_SHORT).show();
                     }
-
-                    {
+                    else{
+                        Toast.makeText(MainActivityCreateAccount.this, "Username or/and password are taken, please try again.", Toast.LENGTH_SHORT).show();
+                    }
+                }
+                else{
                     Toast.makeText(MainActivityCreateAccount.this, "The re-entered password is incorrect.", Toast.LENGTH_SHORT).show();
                 }
             }
-
         });
     }
 }
