+++ CarEshipApp/app/src/main/java/com/example/careshipapp/StaffLoginPage.java
@@ -42,7 +42,7 @@ import androidx.appcompat.app.AppCompatActivity;
                         Toast.makeText(StaffLoginPage.this,"Login successful.",Toast.LENGTH_SHORT).show();
                         successfulLogin();
 
-                    } else if(email.isEmpty() && password.isEmpty()){
+                    } if(email.isEmpty() && password.isEmpty()){
                         Toast.makeText(StaffLoginPage.this,"Add both email and password",Toast.LENGTH_SHORT).show();
                     }else{
                         Toast.makeText(StaffLoginPage.this,"Invalid login attempt.",Toast.LENGTH_SHORT).show();
@@ -52,7 +52,9 @@ import androidx.appcompat.app.AppCompatActivity;
             });
 
             forgetPass = (TextView) findViewById(R.id.forgetPassword);
+
             forgetPass.setOnClickListener(new View.OnClickListener() {
+
                 @Override
                 public void onClick(View view) {
                     Intent intent = new Intent(getApplicationContext(), UpdatePassword.class);
