+++ CarEshipApp/app/src/main/res/layout/activity_main_create_account.xml
@@ -21,7 +21,7 @@
         android:id="@+id/signIn"
         android:layout_width="279dp"
         android:layout_height="56dp"
-        android:layout_marginTop="224dp"
+        android:layout_marginTop="280dp"
         android:gravity="center"
         android:textColor="#736698"
         android:textSize="20sp"
@@ -33,8 +33,8 @@
 
     <EditText
         android:id="@+id/Username"
-        android:layout_width="408dp"
-        android:layout_height="60dp"
+        android:layout_width="410dp"
+        android:layout_height="68dp"
         android:layout_marginTop="16dp"
 
         android:background="#4061DABD"
@@ -51,7 +51,7 @@
     <EditText
         android:id="@+id/CreatePassword"
         android:layout_width="411dp"
-        android:layout_height="66dp"
+        android:layout_height="71dp"
         android:layout_below="@id/Username"
         android:layout_marginTop="24dp"
         android:background="#414083D5"
@@ -70,9 +70,9 @@
     <EditText
         android:id="@+id/passwordReentry"
         android:layout_width="411dp"
-        android:layout_height="60dp"
+        android:layout_height="66dp"
         android:layout_below="@id/Username"
-        android:layout_marginTop="20dp"
+        android:layout_marginTop="24dp"
         android:background="#437454AC"
         android:drawableStart="@drawable/ic_baseline_vpn_key_24"
         android:drawablePadding="20dp"
@@ -91,7 +91,7 @@
         android:layout_width="172dp"
         android:layout_height="53dp"
         android:layout_centerHorizontal="true"
-        android:layout_marginTop="116dp"
+        android:layout_marginTop="24dp"
         android:background="@drawable/button"
         android:backgroundTint="#736698"
         android:text="@string/create_account"
@@ -101,59 +101,12 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/passwordReentry" />
 
-    <EditText
-        android:id="@+id/postAddress"
-        android:layout_width="251dp"
-        android:layout_height="60dp"
-        android:layout_marginTop="16dp"
-        android:background="@color/blue"
-        android:backgroundTint="@color/sky_blue"
-        android:drawableStart="@android:drawable/ic_menu_mylocation"
-        android:drawablePadding="20dp"
-        android:hint="@string/post_address"
-        android:inputType="textPostalAddress"
-        android:padding="20dp"
-        android:textAlignment="center"
-        android:textColor="@color/dark_blue"
-        android:textColorHighlight="@color/dark_blue"
-        android:textColorHint="@color/light_blue"
-        android:textColorLink="@color/dark_blue"
-        android:textSize="20sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/passwordReentry" />
-
-    <EditText
-        android:id="@+id/zipCode"
-        android:layout_width="150dp"
-        android:layout_height="60dp"
-        android:layout_marginStart="4dp"
-        android:layout_marginTop="16dp"
-        android:background="@color/blue"
-        android:backgroundTint="@color/sky_blue"
-        android:drawablePadding="20dp"
-        android:hint="post code"
-        android:inputType="number"
-        android:padding="20dp"
-        android:textAlignment="center"
-        android:textColor="@color/dark_blue"
-        android:textColorHighlight="@color/dark_blue"
-        android:textColorHint="@color/light_blue"
-        android:textColorLink="@color/dark_blue"
-        android:textSize="20sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toEndOf="@+id/postAddress"
-        app:layout_constraintTop_toBottomOf="@+id/passwordReentry" />
-
     <ImageView
         android:id="@+id/imageView14"
-        android:layout_width="298dp"
-        android:layout_height="226dp"
-        android:layout_marginTop="28dp"
+        android:layout_width="371dp"
+        android:layout_height="307dp"
+        android:layout_marginTop="16dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.495"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@drawable/sign_up_form" />
