+++ CarEshipApp/app/src/main/res/layout/update_password.xml
@@ -6,10 +6,10 @@
     android:layout_height="match_parent">
 
     <EditText
-        android:id="@+id/retypeNewPass"
+        android:id="@+id/retypeNewPassword"
         android:layout_width="407dp"
         android:layout_height="74dp"
-        android:layout_marginBottom="24dp"
+        android:layout_marginBottom="36dp"
         android:background="#414083D5"
         android:drawableStart="@drawable/ic_baseline_vpn_key_24"
         android:drawablePadding="20dp"
@@ -18,16 +18,15 @@
         android:inputType="textPassword"
         android:padding="20dp"
         android:textColorHint="@color/bright_blue"
-        app:layout_constraintBottom_toTopOf="@+id/resetBtn"
+        app:layout_constraintBottom_toTopOf="@+id/button2"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent" />
 
     <EditText
-        android:id="@+id/newPass"
+        android:id="@+id/newPassword"
         android:layout_width="411dp"
         android:layout_height="71dp"
-        android:layout_marginBottom="36dp"
+        android:layout_marginBottom="48dp"
         android:background="#414083D5"
         android:drawableStart="@drawable/ic_baseline_vpn_key_24"
         android:drawablePadding="20dp"
@@ -37,7 +36,7 @@
         android:padding="20dp"
         android:textColor="#414083D5"
         android:textColorHint="@color/bright_blue"
-        app:layout_constraintBottom_toTopOf="@+id/retypeNewPass"
+        app:layout_constraintBottom_toTopOf="@+id/retypeNewPassword"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent" />
@@ -46,11 +45,11 @@
         android:id="@+id/textView7"
         android:layout_width="324dp"
         android:layout_height="42dp"
-        android:layout_marginBottom="132dp"
+        android:layout_marginBottom="28dp"
         android:text="Choose your new password"
         android:textColor="@color/bright_blue"
         android:textSize="22sp"
-        app:layout_constraintBottom_toTopOf="@+id/newPass"
+        app:layout_constraintBottom_toTopOf="@+id/newPassword"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.367"
         app:layout_constraintStart_toStartOf="parent" />
@@ -71,44 +70,22 @@
         android:id="@+id/imageView24"
         android:layout_width="347dp"
         android:layout_height="205dp"
-        android:layout_marginBottom="16dp"
+        android:layout_marginBottom="44dp"
         app:layout_constraintBottom_toTopOf="@+id/textView7"
         app:layout_constraintEnd_toEndOf="parent"
-
         app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="1.0"
         app:srcCompat="@drawable/vecteezy_safer_internet_day_" />
 
     <Button
-        android:id="@+id/resetBtn"
+        android:id="@+id/button2"
         android:layout_width="160dp"
         android:layout_height="49dp"
-        android:layout_marginBottom="32dp"
+        android:layout_marginBottom="36dp"
         android:backgroundTint="@color/bright_blue"
         android:text="Reset"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent" />
-
-    <EditText
-        android:id="@+id/EmailAddress"
-        android:layout_width="407dp"
-        android:layout_height="74dp"
-        android:layout_marginTop="17dp"
-        android:layout_marginBottom="12dp"
-        android:background="#414083D5"
-        android:drawableStart="@drawable/ic_baseline_person_outline_24"
-        android:drawablePadding="20dp"
-        android:ems="10"
-        android:hint="@string/type_your_email"
-        android:inputType="textEmailAddress"
-        android:padding="20dp"
-        android:textColor="#414083D5"
-        android:textColorHint="@color/bright_blue"
-        app:layout_constraintBottom_toTopOf="@+id/newPass"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView7"
-        app:layout_constraintVertical_bias="0.233" />
 </androidx.constraintlayout.widget.ConstraintLayout>
