+++ CarEshipApp/app/src/main/res/values/strings.xml
@@ -27,6 +27,4 @@
     <string name="post_address">post address</string>
     <string name="new_password">New password</string>
     <string name="retype_new_password">Retype new password</string>
-    <string name="zip_code">Zip code</string>
-    <string name="type_your_email">Type your email</string>
 </resources>
